+++ app/components/content/GameInstructions.vue
@@ -41,6 +41,48 @@ const error = computed(() => ($colorblindMode.value ? "Orange" : "Red"));
       {{ t("instructions.classic.properties.red_explanation") }}
     </p>
     <p><span>⬆️⬇️</span> {{ t("instructions.classic.properties.arrows") }}</p>
+    <USeparator />
+    <p class="font-roboto text-center text-lg font-bold uppercase">
+      Properties
+    </p>
+    <p>Here is the details of each of the properties columns:</p>
+    <div class="space-y-2">
+      <div class="space-y-1">
+        <p class="font-medium text-(--ui-info)">Gender:</p>
+        <p><span>Possible values:&nbsp;</span> Male, Female or Non-binary</p>
+      </div>
+      <div class="space-y-1">
+        <p class="font-medium text-(--ui-info)">Variant:</p>
+        <p><span>Possible values:&nbsp;</span> Standard, Prime or Umbra</p>
+      </div>
+      <div class="space-y-1">
+        <p class="font-medium text-(--ui-info)">Health:</p>
+        <p>
+          <span>Possible values:&nbsp;</span> Base shield values of Warframes
+          e.g 180, 270, 365
+        </p>
+      </div>
+      <div class="space-y-1">
+        <p class="font-medium text-(--ui-info)">Shield:</p>
+        <p>
+          <span>Possible values:&nbsp;</span> Base shield values of Warframes
+          e.g 0, 135, 180
+        </p>
+      </div>
+      <div class="space-y-1">
+        <p class="font-medium text-(--ui-info)">Progenitor Element:</p>
+        <p>
+          <span>Possible values:&nbsp;</span>Impact, Heat, Cold, Electricity,
+          Toxin, Magnetic or Radiation
+        </p>
+      </div>
+      <div class="space-y-1">
+        <p class="font-medium text-(--ui-info)">Release Year:</p>
+        <p>
+          <span>Possible values:&nbsp;</span>Any year between 2012 and today
+        </p>
+      </div>
+    </div>
     <p class="font-roboto text-center text-lg font-bold uppercase">
       {{ t("instructions.classic.example") }}
     </p>
@@ -57,11 +99,13 @@ const error = computed(() => ($colorblindMode.value ? "Orange" : "Red"));
           properties will appear:
         </p>
       </div>
-      <div class="grid grid-cols-6 gap-1">
-        <ClassicFeedbackRow
-          :guessed-warframe="Inaros!"
-          :correct-warframe="Nezha!"
-        />
+      <div class="overflow-x-auto">
+        <div class="grid w-[160%] grid-cols-7 gap-1">
+          <ClassicFeedbackRow
+            :guessed-warframe="Inaros!"
+            :correct-warframe="Nezha!"
+          />
+        </div>
       </div>
       <div class="space-y-1">
         <p class="font-medium">
@@ -99,11 +143,13 @@ const error = computed(() => ($colorblindMode.value ? "Orange" : "Red"));
         <p>{{ t("instructions.classic.release_year.description") }}</p>
       </div>
       <p>{{ t("instructions.classic.correct_guess") }}</p>
-      <div class="grid grid-cols-6 gap-1">
-        <ClassicFeedbackRow
-          :guessed-warframe="Nezha!"
-          :correct-warframe="Nezha!"
-        />
+      <div class="overflow-x-auto">
+        <div class="grid w-[150%] grid-cols-7 gap-1">
+          <ClassicFeedbackRow
+            :guessed-warframe="Nezha!"
+            :correct-warframe="Nezha!"
+          />
+        </div>
       </div>
     </div>
   </div>
@@ -113,5 +159,10 @@ const error = computed(() => ($colorblindMode.value ? "Orange" : "Red"));
       Guess the Warframe the ability belongs to in {{ defaultAttempts }} tries
     </p>
     <p>{{ t("instructions.ability.subtitle") }}</p>
+    <p>
+      Images begin rotated and are placed in the correct orientation as
+      additional tiles are revealed
+    </p>
+    <p>The order in which tiles are revealed may change between days</p>
   </div>
 </template>