+++ app/components/ui/FeedbackTile.vue
@@ -1,36 +1,37 @@
 <script setup lang="ts">
 const {
-  isCorrect = null,
-  difference,
   fieldLabel,
   fieldValue,
   tooltipDisabled = false,
+  variant = "neutral",
 } = defineProps<{
-  isCorrect?: boolean;
-  difference?: number;
   fieldLabel: string;
   fieldValue?: string | number;
   tooltipDisabled?: boolean;
+  variant?: Result;
 }>();
 
 const tooltipText = computed(() => {
   if (fieldLabel === "Warframe") return `${fieldValue}`;
-  if (isCorrect) return `Correct ${fieldLabel} (${fieldValue})`;
-  if (isCorrect === false) return `Incorrect ${fieldLabel} (${fieldValue})`;
-  return fieldLabel;
+  if (variant === "correct") {
+    return `Correct ${fieldLabel} (${fieldValue})`;
+  } else {
+    return `Incorrect ${fieldLabel} (${fieldValue})`;
+  }
 });
 
+// I might need to make some of these the default styles for all tooltips
 const tooltipStyles = computed(() => {
   const baseStyles = "text-md rounded-none py-2 px-3";
 
-  if (isCorrect) {
+  if (variant === "correct") {
     return {
       content: `${baseStyles} bg-success`,
     };
-  } else if (isCorrect === false) {
-    return { content: `${baseStyles} bg-error` };
+  } else if (variant === "neutral") {
+    return { content: `${baseStyles}` };
   }
-  return { content: `${baseStyles}` };
+  return { content: `${baseStyles} bg-error` };
 });
 </script>
 <template>
@@ -46,21 +47,21 @@ const tooltipStyles = computed(() => {
     <div
       tabindex="0"
       :class="[
-        'relative z-0 min-h-12 w-full border-2 font-semibold text-white shadow-inner transition-colors',
+        'relative z-0 min-h-12 w-full border-2 font-medium break-words text-white shadow-inner transition-colors',
         {
           'border-neutral-500 bg-white/75 dark:border-neutral-700 dark:bg-neutral-900/75':
-            isCorrect === null,
+            variant === 'neutral',
           'border-border-success bg-success hover:shadow-inner hover:brightness-110':
-            isCorrect === true,
+            variant === 'correct',
           'border-border-error bg-error hover:shadow-inner hover:brightness-110':
-            isCorrect === false,
-          'arrow-up': difference && difference > 0,
-          'arrow-down': difference && difference < 0,
+            variant !== 'correct' && variant !== 'neutral',
+          'arrow-up': variant === 'higher',
+          'arrow-down': variant === 'lower',
         },
       ]"
     >
       <div
-        class="relative z-auto flex h-full items-center justify-center pt-2 text-center"
+        class="relative z-auto flex h-full items-center justify-center pt-1 text-center"
       >
         <slot />
       </div>