+++ sentry.client.config.ts
@@ -15,9 +15,6 @@ Sentry.init({
   // sessions when an error occurs.
   replaysOnErrorSampleRate: 1.0,
 
-  // If you don't want to use Session Replay, just remove the line below:
-  integrations: [Sentry.replayIntegration()],
-
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,
 });